chatter: rename id peers: add addr_str property peers: send an announce about self if our pickle changes pinger: rename id prompt: add some extra info
router:
change get_my_address so it reterns a deferred, this is needed for windows because it's adapter isn't ready when we first call.  The deferred is called when the function is finished, then we start making connections.
for get_my_address, in windows we have to re-queue the check for a few seconds to give the adapter time to get ready, then we call the deferred.  This should probably be made more robust by checking a few times and throwing an error if it doesn't "get ready", although it would be nice if there was a better way to check if the adapter is ready.

settings: add ability to specify settings file (not tested)
tuntap: add is_tap/is_tun to windows tuntap
add get_ips functions, need a better way to do this on linux
util: decode_mac wasn't returning in hex